refactor: pull out metrics function preparing the data
